Distinguish setting initializers from setting readers
Setting initializers are intended to be used once during initialization, they
provide complete settings. Among other things, they are expected to set up
rules for each setting so that they may be verified if anything attempts to
alter them in the future.

Setting readers are not currently implemented, but their intended purpose is to
parse setting sources (such as files) to key-value pairs that can be used to
update settings initialized by the setting initializer.