setting: Rework handling to use setting_t
Previously, any setting that was updated through D-Bus was converted to a
string. This was redundant and the sdbus variants are now directly converted
to/from std variants instead.

With the improved type handling, we are also able to keep better track of what
settings were updated. This means the 'SettingsUpdated' signal should only
contain settings where the value changed by the 'Set' call.

This required some rework to the ISettingRule interface, splitting the
conversion and validation into two separate steps. For simplicity, all settings
are now required to obey a rule. If you prefer unconstrained settings, use a
SettingRuleString since that is the most forgiving rule.