Setting: Allow passing string as type for all settings
We already have the logic to convert strings to our internal data types.
Allowing the usage of strings for all settings makes it a lot easier for
clients since they dont need to know what type to use for the 'Set' call.

We still allow the user to pass of the 'correct' type, and rules are still
checked just as if they did.